+++ packages/cli/templates/react/igr-ts/pie-chart/default/files/src/app/__path__/__filePrefix__.tsx
@@ -9,49 +9,48 @@ IgrPieChartModule.register();
 IgrItemLegendModule.register();
 
 const data = [
-    { MarketShare: 30, Company: "Google", },
-    { MarketShare: 15, Company: "Microsoft", },
-    { MarketShare: 30, Company: "Apple", },
-    { MarketShare: 15, Company: "Samsung", },
-    { MarketShare: 10, Company: "Other", },
+  { MarketShare: 30, Company: "Google", },
+  { MarketShare: 15, Company: "Microsoft", },
+  { MarketShare: 30, Company: "Apple", },
+  { MarketShare: 15, Company: "Samsung", },
+  { MarketShare: 10, Company: "Other", },
 ];
 
 export default function $(ClassName)() {
-	const title = 'Pie Chart';
-    const [chartData, setChartData] = useState([]);
-	let legendRef = useRef();
-    let chartRef = useRef();
+  const title = 'Pie Chart';
+  const [chartData, setChartData] = useState([]);
+  let legendRef = useRef();
+  let chartRef = useRef();
 
-	useEffect(() => {
-		setChartData(data);
-	}, []);
+  useEffect(() => {
+    setChartData(data);
+  }, []);
 
-	return (
-		<div>
-			<h1 className={style.title}>{title}</h1>
-			<div>
-				Read more on the&nbsp;
-				<a href="https://www.infragistics.com/products/ignite-ui-react/react/components/piechart.html">
-					official documentation page
-				</a>
-			</div>
-			<div className={style.container}>
-				<div className={style.legend}>
-					<IgrItemLegend ref={legendRef} />
-				</div>
-				<div className={style.chart}>
-					<IgrPieChart dataSource={chartData}
-						labelMemberPath="Company"
-						valueMemberPath="MarketShare"
-						width="500px"
-						height="500px"
-						ref={chartRef}
-						legendLabelMemberPath="Label"
-						legend={legendRef.current}
-						/>
-
-				</div>
-			</div>
-		</div>
-	)
+  return (
+    <div>
+      <h1 className={style.title}>{title}</h1>
+      <div>
+        Read more on the&nbsp;
+        <a href="https://www.infragistics.com/products/ignite-ui-react/react/components/piechart.html">
+          official documentation page
+        </a>
+      </div>
+      <div className={style.container}>
+        <div className={style.legend}>
+          <IgrItemLegend ref={legendRef} />
+        </div>
+        <div className={style.chart}>
+          <IgrPieChart dataSource={chartData}
+            labelMemberPath="Company"
+            valueMemberPath="MarketShare"
+            width="500px"
+            height="500px"
+            ref={chartRef}
+            legendLabelMemberPath="Label"
+            legend={legendRef.current}
+          />
+        </div>
+      </div>
+    </div>
+  )
 }