+++ packages/cli/templates/react/igr-ts/projects/_base/files/src/app/app.test.tsx
@@ -1,8 +1,9 @@
 import { expect, test } from 'vitest';
 import { render } from '@testing-library/react';
 import App from './app';
+import 'element-internals-polyfill';
 
 test('renders without crashing', () => {
-    const wrapper = render(<App />);
-    expect(wrapper).toBeTruthy();
+  const wrapper = render(<App />);
+  expect(wrapper).toBeTruthy();
 });