+++ packages/cli/templates/react/igr-ts/projects/_base/files/src/app/style-utils.ts
@@ -2,11 +2,11 @@
  * Creates a class transformer function that maps matches from a module object
  * @param styles Object map of classes and their scoped name, normally from a CSS Module import
  */
-export default function createClassTransformer(styles: any) {
-	return (classes: string) => {
-		return classes
-			.split(' ')
-			.map(className => styles[className] || className)
-			.join(' ');
-	}
+export default function createClassTransformer(styles: CSSModuleClasses) {
+  return (classes: string) => {
+    return classes
+      .split(' ')
+      .map(className => styles[className] || className)
+      .join(' ');
+  }
 }