+++ packages/igx-templates/index.ts
@@ -1,7 +1,6 @@
 import * as standalone from "./igx-ts";
 import * as legacy from "./igx-ts-legacy";
 
-export * from "./package-resolve";
 export * from "./IgniteUIForAngularTemplate";
 export * from "./AngularTypeScriptFileUpdate";
 export * from "./Update";